+++ internal/database/database.go
@@ -71,8 +71,10 @@ func (dm *Manager) IsChatTrusted(chatID int64) bool {
 func (dm *Manager) GetGlobalChatOverride() (ChatOverride, error) {
 	var chatOverride ChatOverride
 	result := dm.db.Where("chat_id IS NULL").First(&chatOverride)
-	if result.Error != nil && !errors.Is(result.Error, gorm.ErrRecordNotFound) {
-		return chatOverride, result.Error
+	if errors.Is(result.Error, gorm.ErrRecordNotFound) {
+		return ChatOverride{}, nil
+	} else if result.Error != nil {
+		return ChatOverride{}, result.Error
 	}
 	return chatOverride, nil
 }
@@ -80,8 +82,8 @@ func (dm *Manager) GetGlobalChatOverride() (ChatOverride, error) {
 func (dm *Manager) GetChatOverride(chatID int64) (ChatOverride, error) {
 	// Get the default chat override
 	globalChatOverride, err := dm.GetGlobalChatOverride()
-	if err != nil && !errors.Is(err, gorm.ErrRecordNotFound) {
-		return globalChatOverride, err
+	if err != nil {
+		return ChatOverride{}, err
 	}
 
 	// Get the chat-specific override
@@ -90,7 +92,7 @@ func (dm *Manager) GetChatOverride(chatID int64) (ChatOverride, error) {
 	if errors.Is(result.Error, gorm.ErrRecordNotFound) {
 		return globalChatOverride, nil
 	} else if result.Error != nil {
-		return chatOverride, result.Error
+		return ChatOverride{}, result.Error
 	}
 
 	// Merge non-empty fields from chatOverride into globalChatOverride